requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0
streamlit==1.31.0
python-dotenv==1.0.0
pypdf==3.17.4
python-docx==1.1.0
unstructured==0.11.8
unstructured[pdf]==0.11.8
unstructured[local-inference]==0.11.8
pdf2image==1.16.3
pillow==10.2.0
pytesseract==0.3.10
ragas==0.1.4
datasets==2.16.1
matplotlib==3.8.2
seaborn==0.13.1
numpy==1.26.3
opencv-python==4.9.0.80